Use label for Output instead of TreeView
TreeView looks better, but it doesn't let the user select the text, which is a
major problem for the intended use case.

I want to use some other container that let me have nice columns, but in the end
they didn't feel right. So I stuck with a label, which will work just fine.